marta_bot-master/requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
aiopg==1.3.3
aioschedule==0.5.2
asgiref==3.4.1
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.2.0
Babel==2.9.1
black==21.9b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.7
django-import-export==2.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.3
loguru==0.5.3
MarkupPy==1.14
MarkupSafe==2.0.1
motor==2.5.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.2
pymongo==3.12.1
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.1
PyYAML==6.0
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
view==0.1
whitenoise==5.3.0
win32-setctime==1.0.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zipp==3.6.0
